+++ Changelog.md
@@ -8,6 +8,7 @@
 - [EKS 1.28] autoscaling/v2beta2 API version of HorizontalPodAutoscaler is no longer served
 - Update kube-state-metrics cluster tool to v2.10.1
 - Update kubectl to 1.27.10 for EKS 1.28
+- Update cluster-autoscaler v1.28.0 or higher for eks 1.28
 - Fix the "pingfederate-cluster" service label selector
 - [EKS 1.28] service.alpha.kubernetes.io/tolerate-unready-endpoints (deprecated)
 - [EKS 1.28]  Freeze k8s.gcr.io image registry
@@ -27,6 +28,7 @@ _Changes:_
 - [X] PDO-6717 [EKS 1.28] service.alpha.kubernetes.io/tolerate-unready-endpoints (deprecated)
 - [X] PDO-6720 Update kube-state-metrics cluster tool to v2.10.1
 - [X] PDO-6721 Update kubectl to 1.27.10 for EKS 1.28
+- [X] PDO-6723 Update cluster-autoscaler v1.28.0 or higher for eks 1.28
 - [X] PDO-6750 Fix the "pingfederate-cluster" service label selector
 
 v1.18.0.0